+++ src/models/index.js
@@ -1,6 +1,4 @@
 import { otpModel } from "./otp.model.js";
 import { postModel } from "./post.model.js";
-import { userModel } from "./user.model.js";
-import {commentModel} from "./comment.model.js";
-import {voteModel} from "./vote.model.js";
-export { otpModel, postModel, userModel, commentModel, voteModel };
+import { userModel } from "./user.model.js";;
+export { otpModel, postModel, userModel,};
