+++ src/models/post.model.js
@@ -1,5 +1,45 @@
 import mongoose from "mongoose";
 
+const voteSchema = mongoose.Schema(
+  {
+    voteAuthor: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: "User",
+      required: true,
+    },
+    isVote: {
+      type: Boolean,
+      default: false,
+    },
+    voteType: {
+      type: String,
+      enum: ["like", "dislike"],
+      required: true,
+    },
+  },
+  {
+    timestamps: true,
+  }
+);
+
+const commentSchema = mongoose.Schema(
+  {
+    commentAuthor: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: "User",
+      required: true,
+    },
+    comment: {
+      type: String,
+      reqired: true,
+    },
+    votes: [voteSchema],
+  },
+  {
+    timestamps: true,
+  }
+);
+
 const postSchema = mongoose.Schema(
   {
     postAuthor: {
@@ -16,9 +56,7 @@ const postSchema = mongoose.Schema(
       required: true,
     },
 
-    comments: [{ 
-      type: mongoose.Schema.Types.ObjectId,
-       ref: "Comment" }],
+    comments: [commentSchema],
     category: {
       type: String,
       required: true,
